Move warning-related code into its own module
